+++ src/types/result.c
@@ -1,6 +1,6 @@
 /*
 result.c - Une
-Modified 2023-02-10
+Modified 2023-02-12
 */
 
 /* Header-specific includes. */
@@ -21,6 +21,7 @@ const wchar_t *une_result_table[] = {
   L"FLT",
   L"STR",
   L"LIST",
+  L"OBJECT",
   L"FUNCTION",
   L"BUILTIN",
   L"CONTINUE",